requirements.txt
python-docx>=0.8.11
transformers>=4.30.0
openai>=1.0.0
langchain>=0.0.200
pydantic>=2.0.0
loguru>=0.7.0
typer>=0.9.0
pytest>=7.0.0
pytest-cov>=4.1.0
ruff>=0.0.260
mypy>=1.0.0